Trait gomspace_p31u_api::GsEps

pub trait GsEps: Send {
    fn new(bus: &str, addr: u8) -> EpsResult<Self>
    where
        Self: Sized;

    fn ping(&self) -> EpsResult<()>;

    fn reset(&self) -> EpsResult<()>;

    fn reboot(&self) -> EpsResult<()>;

    fn configure_system(&self, config: EpsSystemConfig) -> EpsResult<()>;

    fn configure_battery(&self, config: EpsBatteryConfig) -> EpsResult<()>;

    fn save_battery_config(&self) -> EpsResult<()>;

    fn set_output(&self, channel_mask: u8) -> EpsResult<()>;

    fn set_single_output(
        &self, 
        channel: EpsChannels, 
        value: EpsPowerState, 
        delay: u16
    ) -> EpsResult<()>;

    fn set_input_value(
        &self, 
        in1_voltage: u16, 
        in2_voltage: u16, 
        in3_voltage: u16
    ) -> EpsResult<()>;

    fn set_input_mode(&self, mode: u8) -> EpsResult<()>;

    fn set_heater(
        &self, 
        heater: HeaterSelect, 
        mode: EpsPowerState
    ) -> EpsResult<()>;

    fn reset_system_config(&self) -> EpsResult<()>;

    fn reset_battery_config(&self) -> EpsResult<()>;

    fn reset_counters(&self) -> EpsResult<()>;

    fn get_housekeeping(&self) -> EpsResult<EpsHk>;

    fn get_system_config(&self) -> EpsResult<EpsSystemConfig>;

    fn get_battery_config(&self) -> EpsResult<EpsBatteryConfig>;

    fn get_heater(&self) -> EpsResult<i32>;

    fn watchdog_kick(&self) -> EpsResult<()>;

    fn passthrough(&self, tx: &[u8], rx: &mut [u8]) -> EpsResult<()>;
}

Trait used to represent the GsEps object. Allows for mock objects to be created for unit tests

Required methods

fn new(bus: &str, addr: u8) -> EpsResult<Self> where
Self: Sized,

Construct a new GsEps instance

fn ping(&self) -> EpsResult<()>

Ping subsystems

fn reset(&self) -> EpsResult<()>

Reset

fn reboot(&self) -> EpsResult<()>

Reboot

fn configure_system(&self, config: EpsSystemConfig) -> EpsResult<()>

System configuration

fn configure_battery(&self, config: EpsBatteryConfig) -> EpsResult<()>

Battery configuration

fn save_battery_config(&self) -> EpsResult<()>

Battery configuration save

fn set_output(&self, channel_mask: u8) -> EpsResult<()>

Batch set EPS outputs

fn set_single_output(
&self,
channel: EpsChannels,
value: EpsPowerState,
delay: u16
) -> EpsResult<()>

Set a channel on/off

fn set_input_value(
&self,
in1_voltage: u16,
in2_voltage: u16,
in3_voltage: u16
) -> EpsResult<()>

Set MPPT input level

fn set_input_mode(&self, mode: u8) -> EpsResult<()>

Set the MPPT mode

fn set_heater(&self, heater: HeaterSelect, mode: EpsPowerState) -> EpsResult<()>

Set heater configuration

fn reset_system_config(&self) -> EpsResult<()>

Reset system configuration

fn reset_battery_config(&self) -> EpsResult<()>

Reset battery configuration

fn reset_counters(&self) -> EpsResult<()>

Reset conuters

fn get_housekeeping(&self) -> EpsResult<EpsHk>

Get the telemetry data

fn get_system_config(&self) -> EpsResult<EpsSystemConfig>

Get the system configuration

fn get_battery_config(&self) -> EpsResult<EpsBatteryConfig>

Get battery configuration

fn get_heater(&self) -> EpsResult<i32>

Get the heater status

fn watchdog_kick(&self) -> EpsResult<()>

Kick the hardware watchdog

fn passthrough(&self, tx: &[u8], rx: &mut [u8]) -> EpsResult<()>

Pass a data packet directly through to the device

Loading content...